+++ tabulas-core/src/main/scala/de/tudresden/inf/lat/tabulas/parser/ParserConstant.scala
@@ -9,7 +9,7 @@ object ParserConstant {
   val SpecificationVersion = "1.0.0"
 
   val CommentSymbol: String = "#"
-  val ColonSign: String = ":"
+  val FieldSign: String = ":"
   val EqualsSign: String = "="
   val StandardOrderSign: String = "+"
   val ReverseOrderSign: String = "-"